Fix Chrome Props
