+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './Button'
 export * from './Accordion'
+export * from './TwitterTimelineEmbed'
